add Convolution Node. example : Conv().Input("x").Input("w").Input("b").Output("y").Attr("Kernel", std::vector<int>{2, 2}).Attr("Stride", std::vector<int>{2, 2}) add cfloat include for using FLT_MAX. modify ConstantInit Node Atrribute of Value Type from int to double.

+++ mlfe/unsupported/nodes/initializer_f_cpu.cpp
@@ -10,7 +10,7 @@ namespace mlfe { namespace node {
 template <typename T, typename D = CPUContext>
 struct ConstantInitCpuF : NodeFunctor {
     void Init(OperatorContext *oc) override {
-        _val = oc->attr->GetParam<int>("Value");
+        _val = oc->attr->GetParam<double>("Value");
         _x = oc->inputs[0];
     }
